backend/requirements.txt
Django==4.2.15
psycopg[binary]==3.2.13
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
dj-database-url==2.2.0
celery==5.4.0
redis==5.0.4
python-dotenv==1.0.1

gunicorn==22.0.0

# For type hints and utilities
pytz==2024.1
requests==2.32.3
